requirements.txt
matplotlib
google-generativeai
streamlit
python-dotenv
sympy